requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.5
pydantic==2.5.0
python-dotenv==1.0.0

langchain==0.0.340
langchain-community
llama-index==0.10.61
openai==1.3.0
sentence-transformers==2.2.2

pytest==7.4.3
httpx==0.25.2
